+++ CHANGELOG.md
@@ -5,6 +5,10 @@ of Concierge. Please check the Wiki entry on the release process to understand
 how this file is formatted and how the process works.
 
 ## Unreleased
+### Fixed
+- Avantio: add handling of `check_in_too_near` response code
+- Avantio: mark first MARGIN = 3 days as unavailable
+
 ### Added
 - Ciirus: add new spec for cancel action when `CancelBookingResponse` contains API error